requirements.txt
amqp==2.6.1
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
billiard==3.6.4.0
black==22.6.0
CacheControl==0.12.11
cachetools==4.2.4
celery==4.4.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
class-registry==2.1.2
click==8.0.4
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
dataclasses==0.6
Django==3.2.14
django-braces==1.15.0
django-celery-beat==1.1.1
django-ckeditor==6.3.2
django-debug-toolbar==3.2.4
django-extensions==3.2.0
django-filter==21.1
django-jalali==5.1.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-push-notifications==3.0.0
django-silk==4.2.0
django-widget-tweaks==1.4.11
djangorestframework==3.13.1
docutils==0.19
drf-extra-fields==3.2.1
fcm-django==1.0.9
filters==1.3.2
filters-django==1.0.5
firebase-admin==5.2.0
flit==3.7.1
flit_core==3.7.1
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
gprof2dot==2021.2.21
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.3
incremental==21.3.0
jdatetime==3.8.2
Jinja2==3.0.3
kombu==4.6.11
MarkupSafe==2.0.1
msgpack==1.0.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.30
ndg-httpsclient==0.5.1
ordered-set==4.1.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
proto-plus==1.20.6
protobuf==3.19.4
psycopg2-binary==2.9.3
pusher==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
regex==2022.7.9
requests==2.27.1
rsa==4.8
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
tomli_w==1.0.0
Twisted==22.8.0
txaio==22.2.1
typed-ast==1.5.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.10
vine==1.3.0
zipp==3.6.0
zope.interface==5.4.0
